# v-hormiga-app-main/src/documentation/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MIKNE API",
    "description": "API for MIKNE application",
    "contact": {
      "name": "Support Team",
      "email": "lazcano@icloud.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/DataApp/personacontacto": {
      "get": {
        "tags": [
          "DataApp"
        ],
        "parameters": [
          {
            "name": "cardCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DataApp/direccionfiscal": {
      "get": {
        "tags": [
          "DataApp"
        ],
        "parameters": [
          {
            "name": "cardCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DataApp/direccionentrega": {
      "get": {
        "tags": [
          "DataApp"
        ],
        "parameters": [
          {
            "name": "cardCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}